refactor(tests): split harness routing coverage
