+++ scripts/run-tests.mjs
@@ -885,7 +885,11 @@ export const SUITES = {
     "src/app/api/board/route.test.ts",
     "src/app/api/board/[id]/chat/route.test.ts",
     "src/app/api/sessions/route.test.ts",
-    "src/app/api/chat/send/harness-routing.test.ts",
+    "src/app/api/chat/send/harness-routing-host-session.test.ts",
+    "src/app/api/chat/send/harness-routing-attachments.test.ts",
+    "src/app/api/chat/send/harness-routing-tool-events.test.ts",
+    "src/app/api/chat/send/harness-routing-model-capabilities.test.ts",
+    "src/app/api/chat/send/harness-routing-copilot-jsonl.test.ts",
     "src/app/api/chat/send/offline-queue.test.ts",
     "src/app/api/onboarding/status/route.test.ts",
     "src/app/api/onboarding/install/route.test.ts",
